feat: ESM is used in full
