+++ .eslintrc.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   env: {
     node: true,
     es2021: true
