+++ package.json
@@ -19,11 +19,11 @@
   "main": "packages/index.ts",
   "type": "module",
   "scripts": {
-    "build": "rollup --config rollup.config.mjs",
+    "build": "rollup --config rollup.config.js",
     "build-dts": "tsc && prettier --write ./publish/src",
     "build-strict": "pnpm test-dev && pnpm build && pnpm test-pro",
-    "start": "rollup --config script/start.mjs",
-    "start-server": "rollup --watch --config script/server.mjs",
+    "start": "rollup --config script/start.js",
+    "start-server": "rollup --watch --config script/server.js",
     "test-dev": "jest test/environment/test.ts dev",
     "test-pro": "jest test/environment/test.ts pro",
     "test-crawlPage": "jest test/environment/api/crawlPage.test.ts dev",
